Huge Update!
Massive refactor, memory and performance improvement! Cleaned up a lot
of unnecessary code, classes which were too cluttered. I’m proud to say
that running idle, the engine uses less memory than a Chrome browser
window with medium load (3-4 tabs)! The Windows Task Manager reports
that on average, the engine running idle, uses anywhere from 40mb-130mb
of memory; just launching the application it will start using between
60-90mb, and will gradually decrease.

In terms of performance and memory improvements, the Engine has been
redone to use the float datatype instead of the double datatype. The
main reason for this change is that doubles require twice the amount of
memory to store them; this change means that computations for a
particles movement is quicker, especially in large amounts. The thinking
colors method has also been overhauled. Instead of the runtime needing
to deal with a ton of logic trying to determine if the particles x or y
is between set bounds for velocity, it now has a clean and concise
function which uses the velocity and constricts it to the bounds of the
thinking colors array length.

Removed some classes which could be condensed into a subclass of another
class. Other minor house keeping things include trying to optimize class
structure, creating heap variables and reusing them, renaming classes to
more logical names given what they are aimed to do, reducing logically
heavy functions to math functions for deterministic behavior, minimizing
type casting, limiting creating objects within the main loop.
A Helper utilitiy was added for creation of GUI windows off the main
thread. Usually when a new GUI window is created, if it takes a
noticeable amount of time for it to display, the main window will lock
and you won’t be able to use it until the GUI which was requested has
opened. To get around this, I created a SwingWorker extension class
UIThread. This class’s constructor takes a runnable and then runs the
runnable within a new SwingWorker object. The advantage of this is that
the SwingWorker object execute in a new Thread apart from the main
thread, which gives you the ability to do everything while the GUI
window is created.

I also got rid of the old notification system! Creating and displaying a
notification such as changing the mode of the engine would need to spawn
a new Thread, which is not efficient or practical, especially
considering it’s only a notification. After the draw() method in the
simulation loop, there’s a call to a method named
“headsUpNotifications()”. This method will render a string in the middle
of the window only if a notification has been requested.

There are a ton of other improvements but the list would go on, have fun
and discover what the engine has to offer!