Updated Source+Jar
Got rid of classes which were not really needed.

Removed the need of a Hash map for the editor modes and creation modes for the VPhysics editor, and created an interface with enums for each instead.

Reduced the footprint of the Edge and Vertex classes in the Verlet package by removing the need for Vect2 objects and instead each use 4 float variables; two for their current position, and two for their old position.